feat: 일일권 구매 시 포인트 변동 구현

+++ src/main/java/com/example/gymbo_back_end/core/entity/OrderItem.java
@@ -27,7 +27,7 @@ public class OrderItem {
     @JoinColumn(name = "daily_ticket_seq")
     private DailyTicket dailyTicket;
 
-    private int orderPrice; //주문 가격
+    private Long orderPrice; //주문 가격
 
     private int count; //주문 수량
 
@@ -35,7 +35,7 @@ public class OrderItem {
     //==생성 메서드==//
     public static OrderItem createOrderItem(DailyTicket dailyTicket, int count) {
         String dailyTicketPrice = dailyTicket.getDailyTicketPrice();
-        int price = Integer.parseInt(dailyTicketPrice)*count;
+        long price = Long.parseLong(dailyTicketPrice) * count;
         OrderItem orderItem = OrderItem.builder()
                 .dailyTicket(dailyTicket)
                 .count(count)
